Refactors the admin package
