+++ harvester/core/admin/search.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+
+class ElasticIndexAdmin(admin.ModelAdmin):
+    list_display = ("name", "remote_name", "remote_exists", "error_count", "language", "created_at", "modified_at",
+                    "pushed_at")
